Implement Joining Lobby over Network
Players now receive message that tells them to switch to the lobby screen when they connect to the server.

Lobby State ( Players ) is not yet updated or sent. To do this, preferably change requestHostScene to a PlayerJoined Message and ApplyHostScene to SwitchToLobby message. Other situations ( like switching levels) can rely on sending "Ready" or "SceneReady" messages to the Client.