Generic messages with response and project events
Generic messages with response are now fully implemented.

Project Events have been introduced to artificially schedule execution of methods in another scene without having a bound instance to refer to. For Example lobby calls scene change but level generation in scene context based so it is executed when the event is invoked.